fix audio

+++ AplikasiBelajarAnak/app/src/main/java/com/example/aba/data/database/UploadRecordingResponse.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class UploadRecordingResponse(
 
 	@field:SerializedName("result")
-	val result: Int? = null,
+	val result: Float? = null,
 
 	@field:SerializedName("message")
 	val message: String? = null,
